refactor(frontend): centralize models and types in src/models
