+++ app/frontend/src/models/ChartData.ts
@@ -1,0 +1,6 @@
+export interface ChartDataPoint {
+  date: string;
+  displayDate: string;
+  average: number | null;
+  userAverage: number | null;
+}
