+++ app/frontend/src/models/User.ts
@@ -3,5 +3,5 @@ export interface User {
   email: string;
   name: string;
   avatarUrl?: string;
-  createdAt: string; // ISO string
+  createdAt?: string; // ISO string
 }
